requirements.txt
# MCP Enterprise System Dependencies
# Consolidated requirements for all Python components

# Web Frameworks
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
flask>=2.3.0
flask-cors>=4.0.0

# Async and HTTP
aiohttp>=3.8.0
httpx>=0.24.0
asyncio-mqtt>=0.13.0
websockets>=11.0.0

# Data Validation and Processing
pydantic>=2.0.0
jsonschema>=4.17.0
python-dotenv>=1.0.0
requests>=2.31.0

# Database and Storage
supabase>=1.0.0
psycopg2-binary>=2.9.0
redis>=4.5.0
sqlalchemy>=2.0.0
neo4j>=5.26.0

# Knowledge Graph and AI
graphiti-core[anthropic,google-genai]>=0.1.0
openai>=1.0.0
langwatch>=0.2.0

# LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-google-genai>=2.0.0

# Security and Authentication
cryptography>=41.0.0
pyjwt>=2.8.0
bcrypt>=4.0.0
passlib>=1.7.4

# Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Monitoring and Logging
prometheus-client>=0.17.0
structlog>=23.1.0
sentry-sdk>=1.29.0

# Development and Linting Tools
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
pylint>=2.17.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
click>=8.1.0
colorama>=0.4.6
tqdm>=4.65.0

# Optional: Voice System Dependencies (commented out to avoid conflicts in CI)
# openai-whisper>=20231117
# openai>=1.0.0
# scipy>=1.11.0
# numpy>=1.24.0
# langwatch>=0.2.0

# Optional: LangGraph/LangChain Dependencies
# langchain>=0.1.0
# langgraph>=0.1.0
# langchain-openai>=0.1.0
